+++ Media Player SDK X (crossplatform)/Avalonia/Simple Video Player Avalonia/MainWindow.axaml.cs
@@ -187,7 +187,7 @@ namespace Simple_Video_Player_Avalonia
         {
             var ofd = new OpenFileDialog();
             string[] files = await ofd.ShowAsync(this);
-            if (files?.Length > 0)
+            if (files != null && files.Length > 0)
             {
                 edFilenameOrURL.Text = files[0];
             }