+++ Media Player SDK/WinForms/CSharp/Audio Player/Form1.cs
@@ -6,13 +6,11 @@ namespace Audio_Player_Demo
     using System.Diagnostics;
     using System.IO;
     using System.Windows.Forms;
-
     using VisioForge.Core.MediaPlayer;
-    using VisioForge.Core.UI;
-    using VisioForge.Core.UI.WinForms;
     using VisioForge.Core.Types;
     using VisioForge.Core.Types.Events;
     using VisioForge.Core.Types.MediaPlayer;
+    using VisioForge.Core.UI;
 
     public partial class Form1 : Form
     {
@@ -47,7 +45,8 @@ namespace Audio_Player_Demo
         {
             mmError.Clear();
 
-            MediaPlayer1.FilenamesOrURL.Add(edFilename.Text);
+            MediaPlayer1.Playlist_Clear();
+            MediaPlayer1.Playlist_Add(edFilename.Text);
             MediaPlayer1.Audio_PlayAudio = true;
 
             MediaPlayer1.Source_Mode = MediaPlayerSourceMode.File_DS;