+++ Media Player SDK/WinForms/CSharp/DVD Player/Form1.cs
@@ -8,10 +8,10 @@ namespace DVD_Player_Demo
     using System.Windows.Forms;
     using VisioForge.Core.MediaInfo;
     using VisioForge.Core.MediaPlayer;
-    using VisioForge.Core.UI;
     using VisioForge.Core.Types;
     using VisioForge.Core.Types.Events;
     using VisioForge.Core.Types.MediaPlayer;
+    using VisioForge.Core.UI;
 
     public partial class Form1 : Form
     {
@@ -179,7 +179,8 @@ namespace DVD_Player_Demo
         {
             mmError.Clear();
 
-            MediaPlayer1.FilenamesOrURL.Add(edFilename.Text);
+            MediaPlayer1.Playlist_Clear();
+            MediaPlayer1.Playlist_Add(edFilename.Text);
             MediaPlayer1.Loop = cbLoop.Checked;
             MediaPlayer1.Audio_PlayAudio = true;