+++ Media Player SDK/WinForms/CSharp/Karaoke Demo/Form1.cs
@@ -4,13 +4,12 @@
     using System.Diagnostics;
     using System.IO;
     using System.Windows.Forms;
-
-    using VisioForge.Core.MediaPlayer;
-    using VisioForge.Core.UI;
     using VisioForge.Core.CDG;
+    using VisioForge.Core.MediaPlayer;
     using VisioForge.Core.Types;
     using VisioForge.Core.Types.Events;
     using VisioForge.Core.Types.MediaPlayer;
+    using VisioForge.Core.UI;
 
     public partial class Form1 : Form
     {
@@ -65,7 +64,9 @@
         {
             mmError.Clear();
 
-            MediaPlayer1.FilenamesOrURL.Add(edFilename.Text);
+            MediaPlayer1.Playlist_Clear();
+            MediaPlayer1.Playlist_Add(edFilename.Text);
+
             MediaPlayer1.Audio_PlayAudio = true;
 
             MediaPlayer1.Source_Mode = MediaPlayerSourceMode.File_DS;