+++ Media Player SDK/WinForms/CSharp/Two Windows/Form1.cs
@@ -6,13 +6,11 @@ namespace Two_Windows_Demo
     using System.Diagnostics;
     using System.IO;
     using System.Windows.Forms;
-    using VisioForge.Core;
     using VisioForge.Core.MediaPlayer;
-    using VisioForge.Core.UI;
-    using VisioForge.Core.UI.WinForms;
     using VisioForge.Core.Types;
     using VisioForge.Core.Types.Events;
     using VisioForge.Core.Types.MediaPlayer;
+    using VisioForge.Core.UI;
 
     public partial class Form1 : Form
     {
@@ -77,7 +75,8 @@ namespace Two_Windows_Demo
 
         private async void btStart_Click(object sender, EventArgs e)
         {
-            MediaPlayer1.FilenamesOrURL.Add(edFilename.Text);
+            MediaPlayer1.Playlist_Clear();
+            MediaPlayer1.Playlist_Add(edFilename.Text);
             MediaPlayer1.Loop = cbLoop.Checked;
             MediaPlayer1.Audio_PlayAudio = true;
             MediaPlayer1.Info_UseLibMediaInfo = true;