+++ Media Player SDK/WinForms/CSharp/VR 360 Demo/Form1.cs
@@ -4,13 +4,12 @@
     using System.Diagnostics;
     using System.IO;
     using System.Windows.Forms;
-    using VisioForge.Core;
     using VisioForge.Core.MediaPlayer;
-    using VisioForge.Core.UI;
     using VisioForge.Core.Types;
     using VisioForge.Core.Types.Events;
     using VisioForge.Core.Types.MediaPlayer;
     using VisioForge.Core.Types.VideoEffects;
+    using VisioForge.Core.UI;
 
     public partial class Form1 : Form
     {
@@ -70,7 +69,9 @@
 
             MediaPlayer1.Source_Mode = MediaPlayerSourceMode.LAV;
 
-            MediaPlayer1.FilenamesOrURL.Add(edFilename.Text);
+            MediaPlayer1.Playlist_Clear();
+            MediaPlayer1.Playlist_Add(edFilename.Text);
+
             MediaPlayer1.Audio_PlayAudio = true;
             MediaPlayer1.Info_UseLibMediaInfo = true;
             MediaPlayer1.Audio_OutputDevice = "Default DirectSound Device";