+++ Media Player SDK/WinForms/VB .Net/DVD Player/Form1.vb
@@ -1,15 +1,13 @@
 ' ReSharper disable InconsistentNaming
 
-Imports VisioForge.Core.UI
-Imports VisioForge.Core.MediaInfo
-Imports VisioForge.Core.Types
-Imports VisioForge.Core.UI.WinForms
 Imports System.IO
-Imports VisioForge.Core.Types.MediaInfo
+Imports VisioForge.Core.MediaInfo
 Imports VisioForge.Core.MediaPlayer
+Imports VisioForge.Core.Types
 Imports VisioForge.Core.Types.Events
+Imports VisioForge.Core.Types.MediaInfo
 Imports VisioForge.Core.Types.MediaPlayer
-Imports VisioForge.Core
+Imports VisioForge.Core.UI
 
 Public Class Form1
 
@@ -182,7 +180,8 @@ Public Class Form1
 
         mmError.Clear()
 
-        MediaPlayer1.FilenamesOrURL.Add(edFilename.Text)
+        MediaPlayer1.Playlist_Clear()
+        MediaPlayer1.Playlist_Add(edFilename.Text)
         MediaPlayer1.Loop = cbLoop.Checked
         MediaPlayer1.Audio_PlayAudio = True